+++ nbforager/foragers/ipam.py
@@ -5,19 +5,20 @@ from nbforager.foragers.base_fa import BaseAF
 from nbforager.foragers.forager import Forager
 from nbforager.nb_api import NbApi
 from nbforager.nb_tree import NbTree
+from nbforager.py_tree import PyTree
 
 
 class IpamAF(BaseAF):
     """IPAM Forager."""
 
-    def __init__(self, api: NbApi, root: NbTree, tree: NbTree):
+    def __init__(self, api: NbApi, root: NbTree, tree: NbTree, pynb: PyTree):
         """Init IpamAF.
 
         :param api: NbApi object, connector to Netbox API.
         :param root: NbTree object where raw data from Netbox needs to be saved.
         :param tree: NbTree object where transformed data from Netbox needs to be saved.
         """
-        super().__init__(api, root, tree)
+        super().__init__(api, root, tree, pynb)
         self.aggregates = self.AggregatesF(self)
         self.asn_ranges = self.AsnRangesF(self)
         self.asns = self.AsnsF(self)