+++ nbforager/nb_forager.py
@@ -9,10 +9,11 @@ import logging
 from datetime import datetime
 from pathlib import Path
 
+import pynetbox
+from pynetbox.core.endpoint import Endpoint
 from vhelpers import vstr
 
 from nbforager import nb_tree
-from nbforager.branch.nb_value import NbValue
 from nbforager.foragers.circuits import CircuitsAF
 from nbforager.foragers.core import CoreAF
 from nbforager.foragers.dcim import DcimAF
@@ -27,6 +28,8 @@ from nbforager.messages import Messages
 from nbforager.nb_api import NbApi
 from nbforager.nb_cache import NbCache
 from nbforager.nb_tree import NbTree
+from nbforager.parser.nb_value import NbValue
+from nbforager.py_tree import PyTree
 from nbforager.types_ import LStr, DAny, DiDAny, ODLStr, ODDAny
 
 
@@ -55,7 +58,7 @@ class NbForager:
         sleep: int = 10,
         strict: bool = False,
         # Settings
-        extended_get: bool = False,
+        extended_get: bool = True,
         default_get: ODDAny = None,
         loners: ODLStr = None,
         cache: str = "",
@@ -124,8 +127,9 @@ class NbForager:
 
         :ivar obj root: :py:class:`NbTree` object that holds raw Netbox objects.
             It is data source for the tree.
-        :ivar obj tree: :py:class:`NbTree` object that holds joined Netbox
-            objects.
+        :ivar obj tree: :py:class:`NbTree` object that holds joined Netbox objects.
+        :ivar obj pynb: :py:class:`PyTree` object that holds pynetbox objects,
+            documented on https://pynetbox.readthedocs.io/
         :ivar dict status: Result from Netbox status endpoint. Netbox version.
 
         Application/model foragers:
@@ -162,22 +166,24 @@ class NbForager:
         # data
         self.root: NbTree = NbTree()  # original data
         self.tree: NbTree = NbTree()  # data with joined objects within itself
+        self.pynb: PyTree = PyTree()  # data with pynetbox objects
         self.status: DAny = {}  # updated Netbox status data
 
         self.api = NbApi(**kwargs)
+        self.pyapi = pynetbox.api(url=f"{scheme}://{host}", token=token)
         self.cache: str = make_cache_path(cache, **kwargs)
         self.msgs = Messages(name=self.api.host)
 
         # application foragers
-        self.circuits = CircuitsAF(self.api, self.root, self.tree)
-        self.core = CoreAF(self.api, self.root, self.tree)
-        self.dcim = DcimAF(self.api, self.root, self.tree)
-        self.extras = ExtrasAF(self.api, self.root, self.tree)
-        self.ipam = IpamAF(self.api, self.root, self.tree)
-        self.tenancy = TenancyAF(self.api, self.root, self.tree)
-        self.users = UsersAF(self.api, self.root, self.tree)
-        self.virtualization = VirtualizationAF(self.api, self.root, self.tree)
-        self.wireless = WirelessAF(self.api, self.root, self.tree)
+        self.circuits = CircuitsAF(self.api, self.root, self.tree, self.pynb)
+        self.core = CoreAF(self.api, self.root, self.tree, self.pynb)
+        self.dcim = DcimAF(self.api, self.root, self.tree, self.pynb)
+        self.extras = ExtrasAF(self.api, self.root, self.tree, self.pynb)
+        self.ipam = IpamAF(self.api, self.root, self.tree, self.pynb)
+        self.tenancy = TenancyAF(self.api, self.root, self.tree, self.pynb)
+        self.users = UsersAF(self.api, self.root, self.tree, self.pynb)
+        self.virtualization = VirtualizationAF(self.api, self.root, self.tree, self.pynb)
+        self.wireless = WirelessAF(self.api, self.root, self.tree, self.pynb)
 
     def __repr__(self) -> str:
         """__repr__."""
@@ -253,7 +259,24 @@ class NbForager:
             status = {}
         self.status = status
 
-    def join_tree(self, extra=True) -> NbTree:
+    def join_pynb(self) -> None:
+        """Create self.pynb objects based on self.root dictionaries.
+
+        :return: PyTree object with the pynetbox objects.
+
+        :rtype: PyTree
+        """
+        for app in self.root.apps():  # pylint: disable=R1702
+            for model in getattr(self.root, app).models():
+                objects_d = getattr(getattr(self.root, app), model)
+                for id_, nb_object in objects_d.items():
+                    endpoint: Endpoint = getattr(pynetbox.core.app.App(self.pyapi, app), model)
+                    class_ = endpoint.return_obj
+                    obj = class_(nb_object, self.pyapi, endpoint)
+                    data = getattr(getattr(self.pynb, app), model)
+                    data[id_] = obj
+
+    def join_tree(self, extra=True) -> None:
         """Assemble Netbox objects in NbForager.tree within itself.
 
         The Netbox objects are represented as a multidimensional dictionary.
@@ -271,14 +294,12 @@ class NbForager:
 
         :rtype: NbTree
         """
-        tree = nb_tree.join_tree(self.root)
+        tree: NbTree = nb_tree.join_tree(self.root)
         nb_tree.insert_tree(src=tree, dst=self.tree)
-
         if extra:
             joiner = Joiner(tree=self.tree)
             joiner.join_dcim_devices()
             joiner.join_ipam_ipv4()
-        return self.tree
 
     def read_cache(self) -> None:
         """Read cached data from a pickle file.