+++ nbforager/parser/nb_parser.py
@@ -1,30 +1,78 @@
 # pylint: disable=R0904
 
-"""NbBranch."""
+"""NbParser."""
+from functools import wraps
 from typing import Any, Type, Dict, List
 
 from vhelpers import vstr
 
-from nbforager import wrappers
-from nbforager.exceptions import NbBranchError
-from nbforager.types_ import DAny, SeqStr, LStr, Int, Str
+from nbforager.exceptions import NbParserError
+from nbforager.types_ import DAny, SeqStr, Int, Str
 
 
-class NbBranch:
-    """Extracts a value from a Netbox object using a chain of keys and returns
-    the typed value. Netbox object has None instead of a dictionary when
-    a related object is absent, which is why it is necessary to constantly
-    check the data type. NbBranch returns the desired value with the expected
-    data type, even if the data is missing.
+def check_strict(method):
+    """Wrap method to check value.
+
+    :param method: The method to be decorated.
+
+    :return: The decorated function.
     """
 
-    def __init__(self, data: DAny, strict: bool = False, **kwargs):
-        """Init NbBranch.
+    @wraps(method)
+    def wrapper(self, *args, **kwargs):
+        """Wrap."""
+        result = method(self, *args, **kwargs)
+        if self.strict and not result:
+            keys = "/".join(args)
+            raise NbParserError(f"{keys=} expected in {self._source()}.")
+        return result
+
+    return wrapper
+
+
+def check_in_strict_manner(method):
+    """Wrap method to check value in strict manner, returned value is mandatory.
+
+    :param method: The method to be decorated.
+
+    :return: The decorated function.
+    """
+
+    @wraps(method)
+    def wrapper(self, *args, **kwargs):
+        """Wrap."""
+        strict_actual = self.strict
+        self.strict = True
+
+        result = method(self, *args, **kwargs)
+
+        self.strict = strict_actual
+        if not result:
+            keys = "/".join(args)
+            raise NbParserError(f"{keys=} expected in {self._source()}.")
+
+        return result
+
+    return wrapper
+
+
+class NbParser:
+    """Dictionary parser for extracting values from a Netbox object using a chain of keys.
+
+    Netbox object may have None instead of a dictionary when a related object is absent,
+    requiring constant data type checks. NbParser ensures the desired value is returned
+    with the correct data type, even if the data is missing.
+
+    Raises NbParserError if strict=True and some keys are missing.
+    """
+
+    def __init__(self, data: DAny, strict: bool = False, **kwargs):  # pylint: disable=E0601
+        """Init NbParser.
 
         :param data: Netbox object.
         :type data: dict
 
-        :param strict: True - if data is invalid raise NbBranchError,
+        :param strict: True - if data is invalid raise NbParserError,
             False - if data is invalid return empty data with proper type.
         :type strict: bool
 
@@ -66,8 +114,18 @@ class NbBranch:
         """
         try:
             return self._get_keys(type_=type(None), keys=keys, data=self.data)
-        except NbBranchError:
+        except NbParserError:
             return None
+
+    def bool(self, *keys) -> bool:
+        """Get bool value by keys.
+
+        :param keys: Chaining dictionary keys to retrieve the desired value.
+
+        :return: Boolean value or an empty string if the value is absent.
+        :rtype: bool
+        """
+        return self._get_keys(type_=bool, keys=keys, data=self.data)
 
     def dict(self, *keys) -> Dict:
         """Get dictionary value by keys.
@@ -77,7 +135,7 @@ class NbBranch:
         :return: Dictionary value or an empty dictionary if the value is absent.
         :rtype: dict
 
-        :raise NbBranchError: If strict=True and the value is not a dictionary or key is absent.
+        :raise NbParserError: If strict=True and the value is not a dictionary or key is absent.
         """
         return self._get_keys(type_=dict, keys=keys, data=self.data)
 
@@ -89,7 +147,7 @@ class NbBranch:
         :return: Integer value or 0 if the value is absent.
         :rtype: int
 
-        :raise NbBranchError: If strict=True and the value is not a digit or key is absent.
+        :raise NbParserError: If strict=True and the value is not a digit or key is absent.
         """
         data = self.data
         try:
@@ -98,7 +156,7 @@ class NbBranch:
         except (KeyError, TypeError) as ex:
             if self.strict:
                 type_ = type(ex).__name__
-                raise NbBranchError(f"{type_}: {ex}, {keys=} in {self._source()}") from ex
+                raise NbParserError(f"{type_}: {ex}, {keys=} in {self._source()}") from ex
             return 0
 
         if isinstance(data, int):
@@ -108,7 +166,7 @@ class NbBranch:
             return int(data)
 
         if self.strict:
-            raise NbBranchError(f"{keys=} {int} expected in {self._source()}.")
+            raise NbParserError(f"{keys=} {int} expected in {self._source()}.")
         return 0
 
     def list(self, *keys) -> List:
@@ -119,7 +177,7 @@ class NbBranch:
         :return: List value or an empty list if the value is absent.
         :rtype: list
 
-        :raise NbBranchError: If strict=True and the value is not a list or key is absent.
+        :raise NbParserError: If strict=True and the value is not a list or key is absent.
         """
         return self._get_keys(type_=list, keys=keys, data=self.data)
 
@@ -131,13 +189,13 @@ class NbBranch:
         :return: String value or an empty string if the value is absent.
         :rtype: str
 
-        :raise NbBranchError: If strict=True and the value is not a string or key is absent.
+        :raise NbParserError: If strict=True and the value is not a string or key is absent.
         """
         return self._get_keys(type_=str, keys=keys, data=self.data)
 
     # ======================== strict get methods ========================
 
-    @wrappers.strict_value
+    @check_in_strict_manner
     def strict_dict(self, *keys) -> Dict:
         """Get dictionary value by keys in strict manner, value is mandatory.
 
@@ -147,11 +205,11 @@ class NbBranch:
         :return: Dictionary value.
         :rtype: dict
 
-        :raise NbBranchError: If the value is not a dictionary or key is absent or value is empty.
+        :raise NbParserError: If the value is not a dictionary or key is absent or value is empty.
         """
         return self.dict(*keys)
 
-    @wrappers.strict_value
+    @check_in_strict_manner
     def strict_int(self, *keys) -> Int:
         """Get integer value by keys in strict manner, value is mandatory.
 
@@ -161,11 +219,11 @@ class NbBranch:
         :return: Integer value.
         :rtype: int
 
-        :raise NbBranchError: If the value is not int or key is absent or value is 0.
+        :raise NbParserError: If the value is not int or key is absent or value is 0.
         """
         return self.int(*keys)
 
-    @wrappers.strict_value
+    @check_in_strict_manner
     def strict_list(self, *keys) -> List:
         """Get string value by keys in strict manner, value is mandatory.
 
@@ -175,11 +233,11 @@ class NbBranch:
         :return: List value.
         :rtype: list
 
-        :raise NbBranchError: If the value is not a list or key is absent or value is empty.
+        :raise NbParserError: If the value is not a list or key is absent or value is empty.
         """
         return self.list(*keys)
 
-    @wrappers.strict_value
+    @check_in_strict_manner
     def strict_str(self, *keys) -> Str:
         """Get string value by keys in strict manner, value is mandatory.
 
@@ -188,7 +246,7 @@ class NbBranch:
         :return: String value.
         :rtype: str
 
-        :raise NbBranchError: If the value is not a string or key is absent or value is absent.
+        :raise NbParserError: If the value is not a string or key is absent or value is absent.
         """
         return self.str(*keys)
 
@@ -203,7 +261,7 @@ class NbBranch:
 
         :return: Value with proper data type.
 
-        :raise NbBranchError: If strict=True and key absent or type not match.
+        :raise NbParserError: If strict=True and key absent or type not match.
         """
         try:
             for key in keys:
@@ -211,7 +269,7 @@ class NbBranch:
         except (KeyError, TypeError) as ex:
             if self.strict:
                 ex_type = type(ex).__name__
-                raise NbBranchError(f"{ex_type}: {ex}, {keys=} in {self._source()}.") from ex
+                raise NbParserError(f"{ex_type}: {ex}, {keys=} in {self._source()}.") from ex
             return type_()
 
         if type_ is type(None):
@@ -219,29 +277,13 @@ class NbBranch:
         if not isinstance(data, type_):
             if self.strict:
                 ex_type = "TypeError"
-                raise NbBranchError(f"{ex_type}: {keys=} {type_} expected in {self._source()}.")
+                raise NbParserError(f"{ex_type}: {keys=} {type_} expected in {self._source()}.")
             return type_()
 
         return data
 
-    def tags(self) -> LStr:
-        """Get tag slugs from the data.
-
-        :return: Slugs of tag.
-        :rtype: List[str]
-        """
-        tags_ = self.list("tags")
-        if not tags_:
-            return []
-
-        tags: LStr = []
-        for tag_d in tags_:
-            if tag := self._get_keys(type_=str, keys=["slug"], data=tag_d):
-                tags.append(tag)
-        return tags
-
     def _source(self) -> Str:
-        """Return URL of source object or data."""
+        """Return URL or dictionary of source object."""
         if isinstance(self.data, dict):
             if url := self.data.get("url"):
                 return str(url)