+++ tests/api_/test__nb_api.py
@@ -1,5 +1,3 @@
-# pylint: disable=W0212,R0801,W0621
-
 """Unittests nb_pai.py."""
 import inspect